+++ migrations/000047_group_project_creation.down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE groups DROP COLUMN IF EXISTS can_create_projects;